Renamed libs 'test' folder in Fabric Interop CC to 'testutils'
