+++ core/network/fabric-interop-cc/libs/testutils/mocks/chaincodestub.go
