+++ core/network/fabric-interop-cc/libs/testutils/mocks/clientIdentity.go
