+++ core/network/fabric-interop-cc/libs/testutils/mocks/statequeryiterator.go
