+++ core/network/fabric-interop-cc/libs/testutils/setup.go
@@ -7,7 +7,7 @@
 // test contains the stubs required for counterfeiter to generate the mocks used for tests
 //
 // The mocks are created by running `go generate ./...` as described in the counterfeiter readme
-package test
+package testutils
 
 import (
 	"os"
